requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
streamlit>=1.28.0
PyPDF2>=3.0.0
python-docx>=1.1.0
Pillow>=10.0.0
numpy>=1.24.0
faiss-cpu>=1.7.4

